# Scheduler-integrated/lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: NexusDesk — NITK ECE Management System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /courses:
    get:
      operationId: listCourses
      tags: [courses]
      summary: List courses with attendance stats
      parameters:
        - name: semesterId
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Courses with stats
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CourseWithStats"
    post:
      operationId: createCourse
      tags: [courses]
      summary: Create a course
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CourseInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Course"
  /courses/{courseId}:
    get:
      operationId: getCourse
      tags: [courses]
      summary: Get a course with full stats
      parameters:
        - name: courseId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Course with stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CourseWithStats"
        "404":
          description: Not found
    patch:
      operationId: updateCourse
      tags: [courses]
      summary: Update a course
      parameters:
        - name: courseId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CourseUpdate"
      responses:
        "200":
          description: Updated course
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Course"
    delete:
      operationId: deleteCourse
      tags: [courses]
      summary: Delete a course
      parameters:
        - name: courseId
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted

  # ─── EVENTS ───────────────────────────────────────────────────────────
...
  /attendance:
    get:
      operationId: listAttendance
      tags: [attendance]
      summary: Get attendance summaries per course
      parameters:
        - name: semesterId
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Attendance summaries
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AttendanceSummary"
  /attendance/mark:
    post:
      operationId: markAttendance
      tags: [attendance]
      summary: Mark attendance for an event
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AttendanceInput"
      responses:
        "200":
          description: Marked
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AttendanceRecord"

  # ─── GRADES ───────────────────────────────────────────────────────────
...
  /tasks:
    get:
      operationId: listTasks
      tags: [tasks]
      summary: List tasks, optionally filtered
      parameters:
        - name: category
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Tasks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Task"
    post:
      operationId: createTask
      tags: [tasks]
      summary: Create a task
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TaskInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
  /tasks/{taskId}:
    get:
      operationId: getTask
      tags: [tasks]
      summary: Get a task
      parameters:
        - name: taskId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Task
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
    patch:
      operationId: updateTask
      tags: [tasks]
      summary: Update a task
      parameters:
        - name: taskId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TaskUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
    delete:
      operationId: deleteTask
      tags: [tasks]
      summary: Delete a task
      parameters:
        - name: taskId
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted

  # ─── HARDWARE PROJECTS ────────────────────────────────────────────────
...
components:
  schemas:
...
    Course:
      type: object
      required: [id, subjectCode, name, shortName, creditWeight, minAttendancePct, semesterId, createdAt, updatedAt]
      properties:
        id:
          type: string
        subjectCode:
          type: string
        name:
          type: string
        shortName:
          type: string
        creditWeight:
          type: number
        minAttendancePct:
          type: number
        facultyName:
          type: ["string", "null"]
        roomNumber:
          type: ["string", "null"]
        color:
          type: ["string", "null"]
        semesterId:
          type: string
        createdAt:
          type: string
        updatedAt:
          type: string
    CourseWithStats:
      type: object
      required: [id, subjectCode, name, shortName, creditWeight, minAttendancePct, semesterId, totalClasses, attended, missed, cancelled, effectivePct, isAtRisk, canSkip, mustAttend, createdAt, updatedAt]
      properties:
        id:
          type: string
        subjectCode:
          type: string
        name:
          type: string
        shortName:
          type: string
        creditWeight:
          type: number
        minAttendancePct:
          type: number
        facultyName:
          type: ["string", "null"]
        roomNumber:
          type: ["string", "null"]
        color:
          type: ["string", "null"]
        semesterId:
          type: string
        totalClasses:
          type: integer
        attended:
          type: integer
        missed:
          type: integer
        cancelled:
          type: integer
        effectivePct:
          type: number
        isAtRisk:
          type: boolean
        canSkip:
          type: integer
        mustAttend:
          type: integer
        createdAt:
          type: string
        updatedAt:
          type: string
    CourseInput:
      type: object
      required: [subjectCode, name, shortName, creditWeight, semesterId]
      properties:
        subjectCode:
          type: string
        name:
          type: string
        shortName:
          type: string
        creditWeight:
          type: number
        minAttendancePct:
          type: number
        facultyName:
          type: string
        roomNumber:
          type: string
        color:
          type: string
        semesterId:
          type: string
    CourseUpdate:
      type: object
      properties:
        name:
          type: string
        shortName:
          type: string
        creditWeight:
          type: number
        minAttendancePct:
          type: number
        facultyName:
          type: string
        roomNumber:
          type: string
        color:
          type: string

    # ── Event ──
...
    AttendanceSummary:
      type: object
      required: [courseId, subjectCode, courseName, totalRecords, attended, missed, cancelled, effectivePct, isAtRisk, targetPct]
      properties:
        courseId:
          type: string
        subjectCode:
          type: string
        courseName:
          type: string
        color:
          type: ["string", "null"]
        totalRecords:
          type: integer
        attended:
          type: integer
        missed:
          type: integer
        cancelled:
          type: integer
        effectivePct:
          type: number
        isAtRisk:
          type: boolean
        targetPct:
          type: number
        canSkip:
          type: integer
        mustAttend:
          type: integer
    AttendanceRecord:
      type: object
      required: [id, status, eventId, courseId, createdAt]
      properties:
        id:
          type: string
        status:
          type: string
        note:
          type: ["string", "null"]
        markedAt:
          type: string
        eventId:
          type: string
        courseId:
          type: string
        createdAt:
          type: string
    AttendanceInput:
      type: object
      required: [eventId, courseId, status]
      properties:
        eventId:
          type: string
        courseId:
          type: string
        status:
          type: string
        note:
          type: string

    # ── Grade ──
...
    Task:
      type: object
      required: [id, title, status, category, priority, tags, createdAt, updatedAt]
      properties:
        id:
          type: string
        title:
          type: string
        description:
          type: ["string", "null"]
        dueDate:
          type: ["string", "null"]
        status:
          type: string
        category:
          type: string
        priority:
          type: string
        tags:
          type: array
          items:
            type: string
        linkedCourseId:
          type: ["string", "null"]
        linkedProjectId:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
    TaskInput:
      type: object
      required: [title, category]
      properties:
        title:
          type: string
        description:
          type: string
        dueDate:
          type: string
        category:
          type: string
        priority:
          type: string
        tags:
          type: array
          items:
            type: string
        linkedCourseId:
          type: string
        linkedProjectId:
          type: string
    TaskUpdate:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        dueDate:
          type: string
        status:
          type: string
        category:
          type: string
        priority:
          type: string
        tags:
          type: array
          items:
            type: string

    # ── Project ──
...