# Scheduler-integrated/lib/api-zod/src/generated/types/course.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * NexusDesk — NITK ECE Management System API
 * OpenAPI spec version: 0.1.0
 */

export interface Course {
  id: string;
  subjectCode: string;
  name: string;
  shortName: string;
  creditWeight: number;
  minAttendancePct: number;
  /** @nullable */
  facultyName?: string | null;
  /** @nullable */
  roomNumber?: string | null;
  /** @nullable */
  color?: string | null;
  semesterId: string;
  createdAt: string;
  updatedAt: string;
}