# Scheduler-integrated/lib/api-zod/src/generated/types/courseInput.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * NexusDesk — NITK ECE Management System API
 * OpenAPI spec version: 0.1.0
 */

export interface CourseInput {
  subjectCode: string;
  name: string;
  shortName: string;
  creditWeight: number;
  minAttendancePct?: number;
  facultyName?: string;
  roomNumber?: string;
  color?: string;
  semesterId: string;
}