# Scheduler/lib/api-zod/src/generated/types/attendanceSummary.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * NexusDesk — NITK ECE Management System API
 * OpenAPI spec version: 0.1.0
 */

export interface AttendanceSummary {
  courseId: string;
  subjectCode: string;
  courseName: string;
  /** @nullable */
  color?: string | null;
  totalRecords: number;
  attended: number;
  missed: number;
  cancelled: number;
  effectivePct: number;
  isAtRisk: boolean;
  targetPct: number;
  canSkip?: number;
  mustAttend?: number;
}